【结构指标】新添加的指标没有回显种属编码和种属http://jira.shengtong.com/browse/SAAS-24

+++ src/main/java/cn/staitech/anno/domain/species/Species.java
@@ -24,7 +24,7 @@ public class Species {
      */
     @TableId(value = "species_id", type = IdType.AUTO)
     @ApiModelProperty(value = "种属ID", hidden = true)
-    private Integer speciesId;
+    private Long speciesId;
 
     /**
      * 种属value
